Update RFC changes.

+++ src/models/message.h
@@ -8,7 +8,7 @@
 
 #define IOTA_MESSAGE_ID_BYTES 32
 typedef struct {
-  uint32_t version;                      // The message version. The schema specified in this RFC is for version 1 only.
+  uint8_t version;                       // The message version. The schema specified in this RFC is for version 1 only.
   byte_t trunk[IOTA_MESSAGE_ID_BYTES];   // The Message ID of the first Message we reference.
   byte_t branch[IOTA_MESSAGE_ID_BYTES];  // The Message ID of the second Message we reference.
   size_t payload_length;                 // The length of the Payload.
