+++ src/models/types.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 typedef uint8_t byte_t;
-typedef uint8_t payload_t;
+typedef uint32_t payload_t;
 typedef uint8_t output_t;
 typedef uint8_t input_t;
 typedef uint8_t transaction_t;
